requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.3
pydantic[email]==2.11.7
dnspython==2.7.0
pymongo==4.6.1
python-dotenv==1.0.0 